additional changes

+++ convex/functions.ts
@@ -1,42 +1,65 @@
-import { query, mutation } from "./_generated/server"
+import { query, mutation } from "./_generated/server";
 import { v } from "convex/values";
 
 export const listPrompts = query({
     handler: async (context) => {
-        return await context.db.query("prompts").collect();
-    }
-})
+        const userIdentity = await context.auth.getUserIdentity();
+        
+        if (!userIdentity)
+            throw new Error("User not authenticated!");
+
+        return await context.db.query("prompts").filter((q) => q.eq(q.field("userID"), userIdentity.tokenIdentifier)).collect();
+    },
+});
 
 export const createPrompt = mutation({
     args: {
         title: v.string(),
-        prompt: v.string()
+        prompt: v.string(),
     },
-    handler: async(context, args) => {
+    handler: async (context, args) => {
+        const userIdentity = await context.auth.getUserIdentity();
+
+        if (!userIdentity)
+            throw new Error("User not authenticated!");
+        
         await context.db.insert("prompts", {
             title: args.title,
-            prompt: args.prompt
-        })
-    }
-})
+            prompt: args.prompt,
+            userID: userIdentity.tokenIdentifier,
+        });
+    },
+});
 
 export const deletePrompt = mutation({
     args: {
-        id: v.id("prompts")
+        id: v.id("prompts"),
     },
-    handler: async(context, args) => {
-        await context.db.delete(args.id);
-    }
-})
+    handler: async (context, args) => {
+        const userIdentity = await context.auth.getUserIdentity();
 
-// export const updatePrompt = mutation({
-//     args: {
-//         id: v.id("prompts")
-//     },
-//     handler: async(context, args) => {
-//         await context.db.patch(args.id,{
-//             title: args.title
-//             prompt: args.prompt
-//         });
-//     }
-// })
+        if (!userIdentity)
+            throw new Error("User not authenticated!");
+
+        await context.db.delete(args.id);
+    },
+});
+
+export const updatePrompt = mutation({
+    args: {
+        id: v.id("prompts"),
+        title: v.string(),
+        prompt: v.string(),
+    },
+    handler: async (context, args) => {
+        const userIdentity = await context.auth.getUserIdentity();
+        
+        if (!userIdentity)
+            throw new Error("User not authenticated!");
+
+        await context.db.patch(args.id, {
+            title: args.title,
+            prompt: args.prompt,
+        });
+    },
+});
